container_b/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pika==1.3.2
sqlalchemy==2.0.23
psycopg2-binary
python-dotenv==1.0.0
